rename Reference_mentioned_tag
